Compiler bash script refinement

+++ Part 2/CS2A6/Compiler/car.sh
@@ -1,17 +1,19 @@
 #!/bin/sh
 
-gcc lex.c -o lex
-./lex test.c
-rm lex
+if [ $# = 2 ]
+	then
+		gcc lex.c -o lex
+		./lex $1
+		rm lex
 
-gcc trans.c generator.c -o generator
-./generator
-rm generator
+		gcc trans.c generator.c -o generator
+		./generator
+		rm generator
 
-#cat program.c
+		echo "Code Generation complete"
 
-echo "Code Generation complete"
-
-gcc trans.c program.c -o prog
-./prog
-#rm prog
+		gcc trans.c program.c -o $2
+		./prog
+	else
+		echo "Expected arguments: source desired_exeName"
+fi
